+++ src/org/sosy_lab/cpachecker/core/algorithm/distributed_summaries/distributed_cpa/DssFactory.java
@@ -10,9 +10,11 @@ package org.sosy_lab.cpachecker.core.algorithm.distributed_summaries.distributed
 
 import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
+import com.google.common.collect.ImmutableBiMap;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Maps;
+import java.util.LinkedHashMap;
 import java.util.Map;
 import org.sosy_lab.common.ShutdownNotifier;
 import org.sosy_lab.common.configuration.Configuration;
@@ -50,6 +52,89 @@ import org.sosy_lab.cpachecker.util.predicates.smt.Solver;
 
 public class DssFactory {
 
+  static final class TypeAndLocationCache {
+
+    private static final Map<CFA, ImmutableMap<String, Type>> cachedVariableAndFunctionToTypeMap =
+        new LinkedHashMap<>();
+    private static final Map<CFA, ImmutableBiMap<Integer, CFANode>> integerToNodeMap =
+        new LinkedHashMap<>();
+
+    private TypeAndLocationCache() {}
+
+    static synchronized ImmutableMap<String, Type> getOrCreateTypeMap(
+        CFA pCFA,
+        Configuration pConfiguration,
+        LogManager pLogManager,
+        ShutdownNotifier pShutdownNotifier)
+        throws InvalidConfigurationException, CPATransferException, InterruptedException {
+      if (!cachedVariableAndFunctionToTypeMap.containsKey(pCFA)) {
+        cachedVariableAndFunctionToTypeMap.put(
+            pCFA,
+            ImmutableMap.copyOf(getTypeMap(pCFA, pConfiguration, pLogManager, pShutdownNotifier)));
+      }
+      return cachedVariableAndFunctionToTypeMap.get(pCFA);
+    }
+
+    /**
+     * Get a mapping from variable and function names to their types.
+     *
+     * @param pCfa CFA to get the mapping for
+     * @param pConfiguration configuration to create the solver for the path formula manager
+     * @param pLogManager log manager to create the solver for the path formula manager
+     * @param pShutdownNotifier shutdown notifier to create the solver for the path formula manager
+     * @return a mapping from variable and function names to their types
+     * @throws InvalidConfigurationException if the configuration is invalid for the solver
+     * @throws CPATransferException if the path formula manager cannot create a path formula for the
+     *     given CFA
+     * @throws InterruptedException if the thread is interrupted while creating the path formula
+     */
+    private static Map<String, Type> getTypeMap(
+        CFA pCfa,
+        Configuration pConfiguration,
+        LogManager pLogManager,
+        ShutdownNotifier pShutdownNotifier)
+        throws InvalidConfigurationException, CPATransferException, InterruptedException {
+      try (Solver solver = Solver.create(pConfiguration, pLogManager, pShutdownNotifier)) {
+        PathFormulaManagerImpl pfm =
+            new PathFormulaManagerImpl(
+                solver.getFormulaManager(),
+                pConfiguration,
+                pLogManager,
+                pShutdownNotifier,
+                pCfa,
+                AnalysisDirection.FORWARD);
+        PathFormula pathFormula = pfm.makeEmptyPathFormula();
+        for (CFAEdge edge : pCfa.edges()) {
+          try {
+            pathFormula = pfm.makeAnd(pathFormula, edge);
+          } catch (UnsupportedCodeException e) {
+            // this code might never be executed, so we continue.
+          }
+        }
+        return Maps.toMap(pathFormula.getSsa().allVariables(), pathFormula.getSsa()::getType);
+      }
+    }
+
+    static BiMap<Integer, CFANode> getOrCreateLocationMapping(CFA pCFA) {
+      if (!integerToNodeMap.containsKey(pCFA)) {
+        ImmutableMap<Integer, CFANode> nodeMap =
+            ImmutableMap.copyOf(CFAUtils.getMappingFromNodeIDsToCFANodes(pCFA));
+
+        int minCfaNodeNumber = nodeMap.keySet().stream().min(Integer::compareTo).orElseThrow();
+
+        // All node IDs are shifted such that they start from 0
+        BiMap<Integer, CFANode> cfaNodeIdMap = HashBiMap.create();
+
+        for (Map.Entry<Integer, CFANode> entry : nodeMap.entrySet()) {
+          int index = entry.getKey() - minCfaNodeNumber;
+          cfaNodeIdMap.put(index, entry.getValue());
+        }
+        integerToNodeMap.put(pCFA, ImmutableBiMap.copyOf(cfaNodeIdMap));
+      }
+      return integerToNodeMap.get(pCFA);
+    }
+  }
+
   private DssFactory() {}
 
   /**
@@ -69,10 +154,6 @@ public class DssFactory {
       LogManager pLogManager,
       ShutdownNotifier pShutdownNotifier)
       throws InvalidConfigurationException, CPATransferException, InterruptedException {
-    BiMap<Integer, CFANode> cfaNodeIdMap = createCfaNodeIdMap(pCFA);
-
-    ImmutableMap<String, Type> variableAndFunctionToTypeMap =
-        ImmutableMap.copyOf(getTypeMap(pCFA, pConfiguration, pLogManager, pShutdownNotifier));
     return switch (pCPA) {
       case PredicateCPA predicateCPA ->
           distribute(
@@ -83,9 +164,15 @@ public class DssFactory {
               pOptions,
               pLogManager,
               pShutdownNotifier,
-              cfaNodeIdMap,
-              variableAndFunctionToTypeMap);
-      case CallstackCPA callstackCPA -> distribute(callstackCPA, pBlockNode, pCFA, cfaNodeIdMap);
+              TypeAndLocationCache.getOrCreateLocationMapping(pCFA),
+              TypeAndLocationCache.getOrCreateTypeMap(
+                  pCFA, pConfiguration, pLogManager, pShutdownNotifier));
+      case CallstackCPA callstackCPA ->
+          distribute(
+              callstackCPA,
+              pBlockNode,
+              pCFA,
+              TypeAndLocationCache.getOrCreateLocationMapping(pCFA));
       case FunctionPointerCPA functionPointerCPA -> distribute(functionPointerCPA, pBlockNode);
       case BlockCPA blockCPA -> distribute(blockCPA, pBlockNode, pOptions);
       case ARGCPA argCPA ->
@@ -108,50 +195,11 @@ public class DssFactory {
               pMessageFactory,
               pLogManager,
               pShutdownNotifier);
-      case LocationCPA locationCPA -> distribute(locationCPA, pBlockNode, cfaNodeIdMap);
+      case LocationCPA locationCPA ->
+          distribute(
+              locationCPA, pBlockNode, TypeAndLocationCache.getOrCreateLocationMapping(pCFA));
       case null /*TODO check if null is necessary*/, default -> null;
     };
-  }
-
-  /**
-   * Get a mapping from variable and function names to their types.
-   *
-   * @param pCfa CFA to get the mapping for
-   * @param pConfiguration configuration to create the solver for the path formula manager
-   * @param pLogManager log manager to create the solver for the path formula manager
-   * @param pShutdownNotifier shutdown notifier to create the solver for the path formula manager
-   * @return a mapping from variable and function names to their types
-   * @throws InvalidConfigurationException if the configuration is invalid for the solver
-   * @throws CPATransferException if the path formula manager cannot create a path formula for the
-   *     given CFA
-   * @throws InterruptedException if the thread is interrupted while creating the path formula
-   */
-  private static Map<String, Type> getTypeMap(
-      CFA pCfa,
-      Configuration pConfiguration,
-      LogManager pLogManager,
-      ShutdownNotifier pShutdownNotifier)
-      throws InvalidConfigurationException, CPATransferException, InterruptedException {
-    try (Solver solver = Solver.create(pConfiguration, pLogManager, pShutdownNotifier)) {
-      PathFormulaManagerImpl pfm =
-          new PathFormulaManagerImpl(
-              solver.getFormulaManager(),
-              pConfiguration,
-              pLogManager,
-              pShutdownNotifier,
-              pCfa,
-              AnalysisDirection.FORWARD);
-      PathFormula pathFormula = pfm.makeEmptyPathFormula();
-      for (CFAEdge edge : pCfa.edges()) {
-        try {
-          pathFormula = pfm.makeAnd(pathFormula, edge);
-        } catch (UnsupportedCodeException e) {
-          // this code might never be executed, so we continue.
-        }
-      }
-
-      return Maps.toMap(pathFormula.getSsa().allVariables(), pathFormula.getSsa()::getType);
-    }
   }
 
   private static DistributedConfigurableProgramAnalysis distribute(
@@ -254,21 +302,5 @@ public class DssFactory {
             pMessageFactory,
             pLogManager,
             pShutdownNotifier));
-  }
-
-  static BiMap<Integer, CFANode> createCfaNodeIdMap(CFA pCFA) {
-    ImmutableMap<Integer, CFANode> integerToNodeMap =
-        ImmutableMap.copyOf(CFAUtils.getMappingFromNodeIDsToCFANodes(pCFA));
-
-    int minCfaNodeNumber = integerToNodeMap.keySet().stream().min(Integer::compareTo).orElseThrow();
-
-    // All node IDs are shifted such that they start from 0
-    BiMap<Integer, CFANode> cfaNodeIdMap = HashBiMap.create();
-
-    for (Map.Entry<Integer, CFANode> entry : integerToNodeMap.entrySet()) {
-      int index = entry.getKey() - minCfaNodeNumber;
-      cfaNodeIdMap.put(index, entry.getValue());
-    }
-    return cfaNodeIdMap;
   }
 }