+++ src/org/sosy_lab/cpachecker/core/algorithm/distributed_summaries/distributed_cpa/distributed_block_cpa/BlockViolationConditionOperator.java
@@ -9,8 +9,11 @@
 package org.sosy_lab.cpachecker.core.algorithm.distributed_summaries.distributed_cpa.distributed_block_cpa;
 
 import static org.sosy_lab.common.collect.Collections3.listAndElement;
+import static org.sosy_lab.common.collect.Collections3.transformedImmutableListCopy;
 
+import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.Lists;
 import java.util.List;
 import java.util.Objects;
 import java.util.Optional;
@@ -19,6 +22,7 @@ import org.sosy_lab.cpachecker.core.interfaces.AbstractState;
 import org.sosy_lab.cpachecker.cpa.arg.ARGState;
 import org.sosy_lab.cpachecker.cpa.arg.path.ARGPath;
 import org.sosy_lab.cpachecker.cpa.block.BlockState;
+import org.sosy_lab.cpachecker.cpa.block.BlockTransferRelation;
 import org.sosy_lab.cpachecker.util.AbstractStates;
 
 public class BlockViolationConditionOperator implements ViolationConditionOperator {
@@ -35,8 +39,30 @@ public class BlockViolationConditionOperator implements ViolationConditionOperat
     BlockState topMost =
         Objects.requireNonNull(
             AbstractStates.extractStateByType(pARGPath.getFirstState(), BlockState.class));
+    List<String> previousWitness =
+        pPreviousCondition
+            .map(
+                state ->
+                    Objects.requireNonNull(
+                            AbstractStates.extractStateByType(state, BlockState.class))
+                        .getWitness())
+            .orElse(ImmutableList.of());
+    List<String> currentWitness =
+        ImmutableList.<String>builder()
+            .addAll(previousWitness)
+            .addAll(
+                Lists.reverse(
+                    transformedImmutableListCopy(pARGPath.getFullPath(), BlockTransferRelation::edgeToString)))
+            .build();
     if (!trackHistory) {
-      return Optional.of(topMost);
+      return Optional.of(
+          new BlockState(
+              topMost.getLocationNode(),
+              topMost.getBlockNode(),
+              topMost.getType(),
+              topMost.getViolationConditions(),
+              topMost.getHistory(),
+              currentWitness));
     }
     List<String> previousHistory =
         pPreviousCondition
@@ -52,7 +78,8 @@ public class BlockViolationConditionOperator implements ViolationConditionOperat
             topMost.getBlockNode(),
             topMost.getType(),
             topMost.getViolationConditions(),
-            listAndElement(previousHistory, topMost.getBlockNode().getId()));
+            listAndElement(previousHistory, topMost.getBlockNode().getId()),
+            currentWitness);
     return Optional.of(withHistory);
   }
 }